+++ test/integration/cli-subcommands.test.js
@@ -132,11 +132,19 @@ var subs = [
     ['rbac network-role-tags'],
     ['rbac package-role-tags'],
     ['rbac role-tags'],
+    ['rbac accesskeys'],
+    ['rbac accesskey'],
     ['volume', 'vol'],
     ['volume list', 'volume ls', 'volumes', 'vols'],
     ['volume delete', 'volume rm'],
     ['volume create'],
-    ['volume get']
+    ['volume get'],
+    ['accesskeys'],
+    ['accesskey'],
+    ['accesskey create'],
+    ['accesskey get'],
+    ['accesskey delete', 'accesskey rm'],
+    ['accesskey list', 'accesskey ls']
 ];
 
 // --- Tests