+++ CHANGES.md
@@ -10,6 +10,11 @@ Known issues:
 
 - No changes yet
 
+## 7.18.0
+
+- TRITON-2510 Add `triton accesskeys` and `triton rbac accesskey` commands
+  to manage access keys.
+
 ## 7.17.0
 
 - [node-triton#344] Add docs about tritoncli.ssh.port to instance ssh --help